refactor(commands): extract maintenance commands
